requirements.txt
python-rapidjson
requests
ruamel.yaml
conda
conda-package-handling
conda-smithy
conda-build
ruamel.yaml.jinja2
conda-forge-metadata>=0.11.0
wurlitzer
requests
zstandard
boltons>=23.0.0
py-rattler>=0.21.0,<0.22.0a0
pixi
